Launch CatalogSAML20Fragment from MainFragment instead of catalog fragments
