+++ simplified-main/src/main/java/org/nypl/simplified/main/MainFragmentViewModel.kt
@@ -7,6 +7,8 @@ import io.reactivex.disposables.CompositeDisposable
 import org.librarysimplified.services.api.Services
 import org.nypl.simplified.accounts.api.AccountEvent
 import org.nypl.simplified.accounts.registry.api.AccountProviderRegistryType
+import org.nypl.simplified.books.book_registry.BookRegistryType
+import org.nypl.simplified.books.book_registry.BookStatusEvent
 import org.nypl.simplified.buildconfig.api.BuildConfigurationServiceType
 import org.nypl.simplified.profiles.api.ProfileEvent
 import org.nypl.simplified.profiles.controller.api.ProfilesControllerType
@@ -21,6 +23,8 @@ class MainFragmentViewModel : ViewModel() {
     UnicastWorkSubject.create()
   val profileEvents: UnicastWorkSubject<ProfileEvent> =
     UnicastWorkSubject.create()
+  val registryEvents: UnicastWorkSubject<BookStatusEvent> =
+    UnicastWorkSubject.create()
 
   private val services =
     Services.serviceDirectory()
@@ -28,6 +32,8 @@ class MainFragmentViewModel : ViewModel() {
     services.requireService(ProfilesControllerType::class.java)
   val accountProviders: AccountProviderRegistryType =
     services.requireService(AccountProviderRegistryType::class.java)
+  val bookRegistry: BookRegistryType =
+    services.requireService(BookRegistryType::class.java)
   val buildConfig =
     services.requireService(BuildConfigurationServiceType::class.java)
 
@@ -54,6 +60,17 @@ class MainFragmentViewModel : ViewModel() {
 
   private fun onProfileEvent(event: ProfileEvent) {
     profileEvents.onNext(event)
+  }
+
+  init {
+    this.bookRegistry.bookEvents()
+      .observeOn(AndroidSchedulers.mainThread())
+      .subscribe(this::onBookStatusEvent)
+      .let { subscriptions.add(it) }
+  }
+
+  private fun onBookStatusEvent(event: BookStatusEvent) {
+    registryEvents.onNext(event)
   }
 
   override fun onCleared() {
