+++ simplified-ui-catalog/src/main/java/org/nypl/simplified/ui/catalog/CatalogBookDetailEvent.kt
@@ -3,9 +3,7 @@ package org.nypl.simplified.ui.catalog
 import org.nypl.simplified.accounts.api.AccountID
 import org.nypl.simplified.books.api.Book
 import org.nypl.simplified.books.api.BookFormat
-import org.nypl.simplified.books.api.BookID
 import org.nypl.simplified.ui.errorpage.ErrorPageParameters
-import java.net.URI
 
 sealed class CatalogBookDetailEvent {
 
@@ -15,11 +13,6 @@ sealed class CatalogBookDetailEvent {
 
   data class LoginRequired(
     val account: AccountID
-  ) : CatalogBookDetailEvent()
-
-  data class DownloadWaitingForExternalAuthentication(
-    val bookID: BookID,
-    val downloadURI: URI
   ) : CatalogBookDetailEvent()
 
   data class OpenViewer(
