+++ simplified-ui-catalog/src/main/java/org/nypl/simplified/ui/catalog/CatalogButtons.kt
@@ -53,6 +53,7 @@ class CatalogButtons(
     button.setOnClickListener {
       button.isEnabled = false
       onClick.invoke(button)
+      button.isEnabled = true
     }
     return button
   }
